+++ shopping_list/lib/AppInfo/Application.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OCA\Shopping_List\AppInfo;
 
+use OCA\Shopping_List\Notification\Notifier;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
@@ -17,6 +18,7 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function register(IRegistrationContext $context): void {
+		$context->registerNotifierService(Notifier::class);
 	}
 
 	public function boot(IBootContext $context): void {